chore(): move meta settings actions and clean up actions-file
