+++ tavla/app/(admin)/edit/[id]/components/Settings/actions.ts
@@ -4,32 +4,43 @@ import {
     TFormFeedback,
     getFormFeedbackForError,
 } from 'app/(admin)/utils'
-import { userCanEditBoard } from 'app/(admin)/utils/firebase'
+import {
+    initializeAdminApp,
+    userCanEditBoard,
+    userCanEditOrganization,
+} from 'app/(admin)/utils/firebase'
 import { handleError } from 'app/(admin)/utils/handleError'
 import { getBoard } from 'Board/scenarios/Board/firebase'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
-import { TLocation } from 'types/meta'
-import { TBoard, TBoardID, TTheme } from 'types/settings'
+import { TFontSize, TLocation } from 'types/meta'
 import {
-    moveBoard,
-    saveFont,
-    saveLocation,
-    saveTitle,
-} from '../MetaSettings/actions'
+    TBoard,
+    TBoardID,
+    TFooter,
+    TOrganizationID,
+    TTheme,
+} from 'types/settings'
 import { isRedirectError } from 'next/dist/client/components/redirect-error'
 import { isEmptyOrSpaces, isOnlyWhiteSpace } from 'app/(admin)/edit/utils'
 import { firestore } from 'firebase-admin'
 import * as Sentry from '@sentry/nextjs'
-export async function userHasAccessToEdit(bid: string) {
+import { getUserFromSessionCookie } from 'app/(admin)/utils/server'
+import { getWalkingDistanceTile } from '../../actions'
+
+initializeAdminApp()
+
+async function userHasAccessToEditBoard(bid: string) {
     const access = await userCanEditBoard(bid)
     if (!access) return redirect('/')
 }
+
 export async function saveSettings(data: FormData) {
     const title = data.get('title') as string
     const bid = data.get('bid') as TBoardID
     const viewType = data.get('viewType') as string
     const theme = data.get('theme') as TTheme
+    const font = data.get('font') as TFontSize
 
     let newOrganization = data.get('toOrg') as string | undefined
     const oldOrganization = data.get('fromOrg') as string
@@ -47,6 +58,9 @@ export async function saveSettings(data: FormData) {
     } else {
         location = undefined
     }
+
+    const footer = data.get('footer') as string
+    const override = (data.get('override') as string) === 'on'
 
     const board = await getBoard(bid)
     const errors = {} as Record<InputType, TFormFeedback>
@@ -71,11 +85,11 @@ export async function saveSettings(data: FormData) {
 
         await saveTitle(bid, title)
         await moveBoard(bid, personal, newOrganization, oldOrganization)
-        await saveLocation(bid, location)
-        await saveFont(bid, data)
+        await saveLocation(board, location)
+        await saveFont(bid, font)
         await setTheme(bid, theme)
         await setViewType(board, viewType)
-        await setFooter(bid, data)
+        await setFooter(bid, { footer, override })
 
         revalidatePath(`/edit/${bid}`)
     } catch (error) {
@@ -87,25 +101,23 @@ export async function saveSettings(data: FormData) {
         return errors
     }
 }
-export async function setFooter(bid: TBoardID, data: FormData) {
-    const access = userCanEditBoard(bid)
-    if (!access) return redirect('/')
 
-    const message = data.get('footer') as string
-    const shouldOverrideOrgFooter = (data.get('override') as string) === 'on'
+async function setFooter(bid: TBoardID, { footer, override }: TFooter) {
+    userHasAccessToEditBoard(bid)
 
     let newFooter = {}
 
-    const validFooter =
-        message && !isOnlyWhiteSpace(message) && message.trim() !== ''
+    const footerContainsText =
+        footer && !isOnlyWhiteSpace(footer) && footer.trim() !== ''
 
-    if (validFooter) {
-        newFooter = { footer: message, override: shouldOverrideOrgFooter }
-    } else if (shouldOverrideOrgFooter) {
-        newFooter = { override: shouldOverrideOrgFooter }
+    if (footerContainsText) {
+        newFooter = { footer: footer, override: override }
+    } else if (override) {
+        newFooter = { override: override }
     } else {
         newFooter = firestore.FieldValue.delete()
     }
+
     try {
         await firestore().collection('boards').doc(bid).update({
             footer: newFooter,
@@ -122,8 +134,9 @@ export async function setFooter(bid: TBoardID, data: FormData) {
         return handleError(error)
     }
 }
-export async function setTheme(bid: TBoardID, theme?: TTheme) {
-    userHasAccessToEdit(bid)
+
+async function setTheme(bid: TBoardID, theme?: TTheme) {
+    userHasAccessToEditBoard(bid)
 
     try {
         await firestore()
@@ -146,9 +159,9 @@ export async function setTheme(bid: TBoardID, theme?: TTheme) {
         return handleError(error)
     }
 }
-export async function setViewType(board: TBoard, viewType: string) {
-    const access = await userCanEditBoard(board.id)
-    if (!access) return redirect('/')
+
+async function setViewType(board: TBoard, viewType: string) {
+    userHasAccessToEditBoard(board.id ?? '')
 
     const shouldDeleteCombinedTiles = viewType === 'separate'
 
@@ -166,5 +179,139 @@ export async function setViewType(board: TBoard, viewType: string) {
         revalidatePath(`/edit/${board.id}`)
     } catch (e) {
         handleError(e)
+    }
+}
+
+async function saveTitle(bid: TBoardID, title: string) {
+    userHasAccessToEditBoard(bid)
+
+    try {
+        await firestore()
+            .collection('boards')
+            .doc(bid)
+            .update({
+                'meta.title': title.substring(0, 50),
+                'meta.dateModified': Date.now(),
+            })
+        revalidatePath(`/edit/${bid}`)
+    } catch (error) {
+        Sentry.captureException(error, {
+            extra: {
+                message: 'Error while saving title of board tile',
+                boardID: bid,
+            },
+        })
+        return handleError(error)
+    }
+}
+
+async function saveFont(bid: TBoardID, font: TFontSize) {
+    userHasAccessToEditBoard(bid)
+
+    try {
+        await firestore()
+            .collection('boards')
+            .doc(bid)
+            .update({ 'meta.fontSize': font, 'meta.dateModified': Date.now() })
+        revalidatePath(`/edit/${bid}`)
+    } catch (error) {
+        Sentry.captureException(error, {
+            extra: {
+                message: 'Error while updating font size of board',
+                boardID: bid,
+            },
+        })
+        return handleError(error)
+    }
+}
+
+async function saveLocation(board: TBoard, location?: TLocation) {
+    userHasAccessToEditBoard(board.id ?? '')
+
+    try {
+        await firestore()
+            .collection('boards')
+            .doc(board.id ?? '')
+            .update({
+                tiles: await getTilesWithDistance(board, location),
+                'meta.location': location ?? firestore.FieldValue.delete(),
+                'meta.dateModified': Date.now(),
+            })
+        revalidatePath(`/edit/${board.id}`)
+    } catch (error) {
+        Sentry.captureException(error, {
+            extra: {
+                message: 'Error while updating location of board',
+                boardID: board.id,
+                location: location,
+            },
+        })
+        return handleError(error)
+    }
+}
+
+async function getTilesWithDistance(board: TBoard, location?: TLocation) {
+    return await Promise.all(
+        board.tiles.map(async (tile) => {
+            return await getWalkingDistanceTile(tile, location)
+        }),
+    )
+}
+
+async function moveBoard(
+    bid: TBoardID,
+    personal: boolean,
+    toOrganization: TOrganizationID | undefined,
+    fromOrganization?: TOrganizationID,
+) {
+    const user = await getUserFromSessionCookie()
+    if (!user) return redirect('/')
+
+    userHasAccessToEditBoard(bid)
+
+    if (fromOrganization) {
+        const canEdit = await userCanEditOrganization(fromOrganization)
+        if (!canEdit) return redirect('/')
+    }
+
+    if (toOrganization) {
+        const canEdit = await userCanEditOrganization(toOrganization)
+        if (!canEdit) return redirect('/')
+    }
+
+    try {
+        if (fromOrganization)
+            await firestore()
+                .collection('organizations')
+                .doc(fromOrganization)
+                .update({ boards: firestore.FieldValue.arrayRemove(bid) })
+        else
+            await firestore()
+                .collection('users')
+                .doc(user.uid)
+                .update({ owner: firestore.FieldValue.arrayRemove(bid) })
+
+        if (toOrganization && !personal)
+            await firestore()
+                .collection('organizations')
+                .doc(toOrganization)
+                .update({ boards: firestore.FieldValue.arrayUnion(bid) })
+        else
+            await firestore()
+                .collection('users')
+                .doc(user.uid)
+                .update({ owner: firestore.FieldValue.arrayUnion(bid) })
+
+        revalidatePath(`/edit/${bid}`)
+    } catch (error) {
+        Sentry.captureException(error, {
+            extra: {
+                message: 'Error while moving board to new organization',
+                boardID: bid,
+                newOrg: toOrganization,
+                oldOrg: fromOrganization,
+            },
+        })
+        return handleError(error)
     }
 }
