+++ tavla/app/(admin)/edit/[id]/components/Settings/components/Footer.tsx
