+++ tavla/app/(admin)/edit/[id]/components/Settings/components/ViewType.tsx
