+++ tavla/app/(admin)/folders/components/FontSelect/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { TFontSize } from 'types/meta'
 import { setFontSize as setFontSizeAction } from './actions'
-import { FontChoiceChip } from 'app/(admin)/edit/[id]/components/Settings/FontChoiceChip'
+import { FontChoiceChip } from 'app/(admin)/edit/[id]/components/Settings/components/FontChoiceChip'
 import { SubmitButton } from 'components/Form/SubmitButton'
 import { useToast } from '@entur/alert'
 import { TOrganizationID } from 'types/settings'
